feat: add CODEOWNERS file and dev.yml and sec.yml workflows
Add a CODEOWNERS file to specify that @apoupard is the code owner for all files in the repository. This helps to ensure that the appropriate person is notified and responsible for reviewing and maintaining the code.

Add dev.yml and sec.yml workflows to automate the development and security processes. These workflows are triggered on push events and perform various tasks such as running tests, generating packages, and analyzing code quality and security. The workflows are configured to use specific actions and require certain secrets to be provided for authentication and authorization purposes.