chore(build.gradle.kts): remove unused plugin and update plugin import feat(sonar-project.properties): add SonarQube configuration for Kotlin project The unused plugin "io.komune.fixers.gradle.publish" has been removed from the build.gradle.kts file. Additionally, the plugin import for "io.komune.fixers.gradle.kotlin.jvm" has been updated.
A new file, sonar-project.properties, has been added. This file contains the SonarQube configuration for the Kotlin project. It specifies the organization and project key for SonarQube analysis, as well as the source directories and language for analysis. The detekt.yml file is also specified as the report path for Kotlin detekt analysis.